// src/types/enum/RequestErrorEnum.ts
export enum RequestErrorEnum {
    InvalidRequest = 'invalid_request',
    InvalidRequestType = 'invalid_request_type',
    InvalidPayload = 'invalid_payload',
    InvalidVerificationData = 'invalid_verification_data',
    MissingVerificationData = 'missing_verification_data',
    InvalidSender = 'invalid_sender'
}